requirements.txt
python-dateutil==2.8.2
xmltodict==0.12.0
openpyxl==3.0.7
requests==2.26.0
pymorphy2==0.9.1